+++ go.mod
@@ -4,7 +4,6 @@ go 1.19
 
 require (
 	github.com/bufbuild/connect-go v1.5.1
-	github.com/bufbuild/connect-grpchealth-go v1.0.1-0.20230503165053-2a56ef052862
 	github.com/google/go-cmp v0.5.9
 	google.golang.org/protobuf v1.28.1
 )
